Add files via upload
Radiomic feature extraction provides objective quantitative data that can be analyzed and used to understand brain tumor characteristics and patterns of progression which is often difficult to identify with the naked eye. These features, known as quantitative imaging biomarkers, provide a comprehensive method to identify regions of interest that can be analyzed for disease aggressiveness and thereby help provide for personalized medical care and treatment. With AI to target this research, we intend to use machine learning algorithms to further identify and classify phases in the disease and what characteristics represent those phases.